add thread pool class and modify LTSF to meet the thread pools needs
This commit changes the C++ standard to C++ 14 to make use of std::make_unique. The thread pool fully utilizes the CPU to find BRDF approximations. Since only approximations of the same view direction depend on another a good parallelization is possible. Also implements fitting of the whole LUT of view directions and roughness values.